+++ internal/hybrid/alias.go
@@ -2,12 +2,13 @@ package hybrid
 
 import (
 	"fmt"
+
 	"github.com/kota-yata/byrd-mp3/internal/common"
 )
 
 // cosine value for butterfly calculation
 // aliasReductionCS[i] = 1 / sqrt(1 + a[i]^2) where a[i] is specific coefficients defined in the standard
-var aliasReductionCS = [8]float64{
+var aliasReductionCS = [8]float32{
 	0.857492925712,
 	0.881741997318,
 	0.949628649103,
@@ -20,7 +21,7 @@ var aliasReductionCS = [8]float64{
 
 // sine value for butterfly calculation
 // aliasReductionCA[i] = a[i] / sqrt(1 + a[i]^2) where a[i] is specific coefficients defined in the standard
-var aliasReductionCA = [8]float64{
+var aliasReductionCA = [8]float32{
 	-0.514495755427,
 	-0.471731968565,
 	-0.313377454204,
@@ -31,7 +32,7 @@ var aliasReductionCA = [8]float64{
 	-0.00369997467375,
 }
 
-func ApplyAliasReduction(gc *common.GranuleChannelInfo, values []float64) error {
+func ApplyAliasReduction(gc *common.GranuleChannelInfo, values []float32) error {
 	if gc == nil {
 		return fmt.Errorf("nil granule channel info")
 	}
